Moved utils functions for stein kernels into stein dedicated file.
